+++ packages/ice/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.0.1
 
+- [fix] file resolve when use css module in server compiling
 - [fix] lock version of `@ice/bundles`, `@ice/webpack-config` and `@ice/route-manifest`
 - [feat] support external config for server
 - [feat] redirect imports for data loader