+++ packages/ice/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v3.0.2
+
+- [fix] change main fields order for ssr
+
 ## v3.0.0
 
 - [feat] provider basic service for web framework `ice`